Optimize tests (#246)
* Move fixture to conftest.py

* Use monkeypatch instead of patch
